// shopify-app/mini-project-2/app/constrant/enum.ts
export enum ALL_CODES {
  EARN = "EARN",
  SPEND = "SPEND",
}
export enum EXCHANGE {
  ID = "id",
  POINT_NUMBER = "pointNumber",
  MONEY_AMOUNT = "moneyAmount",
  REDEEMED_CODES = "Redeemed_Code",
  CREATED_AT = "created_at",
  TYPE = "type",
  ALL_CODE = "allCode",
  CONDITION = "condition",
}
export enum RESTAPI {
  PUT = "PUT",
  GET = "GET",
  POST = "POST",
  DELETE = "DELETE",
}

export enum USER {
  ID = "id",
  CUSTOMER_ID = "customerId",
  FIRST_NAME = "firstName",
  LAST_NAME = "lastName",
  EMAIL = "email",
  CREATED_AT = "created_at",
  TOTAL_POINT = "total_point",
}
export enum POINT {
  EARN_POINT = "EARN_POINT",
  SPEND_POINT = "SPEND_POINT",
  TYPE = "TYPE",
  ADD = "ADD",
  UPDATE = "UPDATE",
  DELETE = "DELETE",
  TYPE_ACTIONS = "TYPE_ACTIONS",
  EARN_POINT_DATA = "EARN_POINT_DATA",
  SPEND_POINT_DATA = "SPEND_POINT_DATA",
  ORDER_BY = "orderBy",
  NUMBER_LIMIT = "3",
  NUMBER_LIMIT_LOGS = "5",
  LOG = "LOG",
  GROUP = "GROUP",
  CUSTOMER_ID = "customerId",
  TOTAL_POINTS = "total_points",
  REASON = "reason",
  TIER_NAME = "tierName",
  POINT_RATE = "pointRate",
  MIN_SPENT = "min_spent",
  UPDATE_POINT = "update_point",
  UPDATE_RANK = "update_rank",
  RANKID = "rankId",
  MONEY = "money",
  SYNC_DATA = "sync_data",
}
export enum REDIS_KEY {
  IMAGE = "IMAGE",
}
export enum REVIEW {
  APPROVED = "APPROVED",
  UN_APPROVED = "UN_APPROVED",
  ID_REVIEW = "idReview",
  STATE = "state",
  RATING = "rating",
  CUSTOMER_ID = "customerId",
  ID = "id",
}
export enum Location_Api {
  CHANGE_REVIEW_STATE = "CHANGE-REVIEW-STATE",
  DETAIL_REVIEW = "DETAIL-REVIEW",
  DELETE_REVIEW = "DELETE-REVIEW",
}